feat: modal에 선택된 데이터 전달하기
